atividade-10-padb/requirements.txt
django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
drf_yasg==1.21.11
django_cors_headers==4.9.0